+++ Changelog.md
@@ -6,6 +6,7 @@
 - Can view clients.
 - Can create clients.
 - Can delete client.
+- Can put update client.
 - Can bulk delete clients.
 - Can order data in table service and client.
 - Filter client dialog.